feat: add e2e test

+++ src/task/service/DeleteTaskService.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from "tsyringe";
+import type { ProjectRepository } from "@/project/infra";
 import type { EntityId } from "@/shared/domain/Entity";
 import { NotFoundError } from "@/shared/Errors";
 import type { TaskRepository } from "@/task/infra";
@@ -13,6 +14,8 @@ export type DeleteTaskServiceResponse = undefined;
 export class DeleteTaskService {
 	constructor(
 		@inject("TaskRepository") private readonly taskRepository: TaskRepository,
+		@inject("ProjectRepository")
+		private readonly projectRepository: ProjectRepository,
 	) {}
 
 	async execute(
@@ -22,6 +25,17 @@ export class DeleteTaskService {
 
 		if (!task) throw NotFoundError.task(params.id);
 
+		const project = await this.projectRepository.findById(task.projectId);
+
+		if (!project) throw NotFoundError.project(task.projectId);
+
+		const updatedTaskIds = project.taskIds.filter(
+			(taskId) => taskId !== params.id,
+		);
+
+		project.updateTaskIds(updatedTaskIds);
+
+		await this.projectRepository.save(project);
 		await this.taskRepository.delete(params.id);
 	}
 }
